Remove TF
