+++ .github/workflows/python-package.yml
@@ -38,5 +38,5 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        python -m pip install scikit-learn tensorflow tensorflow_probability
+        python -m pip install scikit-learn
         pytest
