+++ src/copairs/map.py
@@ -5,8 +5,7 @@ import numpy as np
 import pandas as pd
 from statsmodels.stats.multitest import multipletests
 
-from copairs.compute import cosine_indexed
-from copairs import compute_np
+from copairs import compute
 from copairs.matching import Matcher, MatcherMultilabel
 
 logger = logging.getLogger('copairs')
@@ -96,19 +95,19 @@ def run_pipeline(
                             count=total_neg)
 
     logger.info('Computing positive similarities...')
-    pos_dists = cosine_indexed(feats, pos_pairs, batch_size).astype(np.float32)
+    pos_dists = compute.pairwise_cosine(feats, pos_pairs, batch_size)
 
     logger.info('Computing negative similarities...')
-    neg_dists = cosine_indexed(feats, neg_pairs, batch_size).astype(np.float32)
+    neg_dists = compute.pairwise_cosine(feats, neg_pairs, batch_size)
 
     logger.info('Building rank lists...')
     rel_k_list, counts = build_rank_lists(pos_pairs, neg_pairs, pos_dists, neg_dists)
 
     logger.info('Computing average precision...')
-    ap_scores, null_confs = compute_np.compute_ap_contiguos(rel_k_list, counts)
+    ap_scores, null_confs = compute.compute_ap_contiguos(rel_k_list, counts)
 
     logger.info('Computing p-values...')
-    p_values = compute_np.compute_p_values(ap_scores, null_confs, null_size, seed=seed)
+    p_values = compute.compute_p_values(ap_scores, null_confs, null_size, seed=seed)
 
     logger.info('Creating result DataFrame...')
     meta['average_precision'] = ap_scores
@@ -126,14 +125,14 @@ def create_neg_query_solver(neg_pairs, neg_dists):
     neg_dists = neg_dists[sort_ix]
 
     neg_ix, neg_counts = np.unique(neg_ix, return_counts=True)
-    neg_cutoffs = compute_np.to_cutoffs(neg_counts)
+    neg_cutoffs = compute.to_cutoffs(neg_counts)
 
     def negs_for(query: np.ndarray):
         locs = np.searchsorted(neg_ix, query)
         sizes = neg_counts[locs]
         start = neg_cutoffs[locs]
         end = start + sizes
-        slices = compute_np.concat_ranges(start, end)
+        slices = compute.concat_ranges(start, end)
         batch_dists = neg_dists[slices]
         return batch_dists, sizes
     return negs_for
@@ -158,8 +157,8 @@ def build_rank_lists_multi(pos_pairs, pos_dists, pos_counts, negs_for, null_size
         ix_sort = np.lexsort([1 - dist_all, ix])
         rel_k_list = labels[ix_sort]
         _, counts = np.unique(ix, return_counts=True)
-        ap_scores, null_confs = compute_np.compute_ap_contiguos(rel_k_list, counts)
-        p_values = compute_np.compute_p_values(ap_scores, null_confs, null_size, seed=seed)
+        ap_scores, null_confs = compute.compute_ap_contiguos(rel_k_list, counts)
+        p_values = compute.compute_p_values(ap_scores, null_confs, null_size, seed=seed)
 
         ap_scores_list.append(ap_scores)
         p_values_list.append(p_values)
@@ -205,10 +204,10 @@ def run_pipeline_multilabel(
     neg_pairs = np.unique(neg_pairs, axis=0)
 
     logger.info('Computing positive similarities...')
-    pos_dists = cosine_indexed(feats, pos_pairs, batch_size).astype(np.float32)
+    pos_dists = compute.pairwise_cosine(feats, pos_pairs, batch_size)
 
     logger.info('Computing negative similarities...')
-    neg_dists = cosine_indexed(feats, neg_pairs, batch_size).astype(np.float32)
+    neg_dists = compute.pairwise_cosine(feats, neg_pairs, batch_size)
 
     logger.info('Computing mAP and p-values per label...')
     negs_for = create_neg_query_solver(neg_pairs, neg_dists)
