+++ src/copairs/replicating.py
@@ -4,14 +4,14 @@ from typing import List, Literal
 import numpy as np
 import pandas as pd
 
-from copairs.compute import corrcoef_indexed
+from copairs.compute import pairwise_corr
 from .matching import Matcher
 
 
 def corr_from_null_pairs(X: np.ndarray, null_pairs, n_replicates):
     """ Correlation from a given list of unnamed pairs.  """
     null_pairs = np.asarray(null_pairs, int)
-    corrs = corrcoef_indexed(X, null_pairs, batch_size=20000)
+    corrs = pairwise_corr(X, null_pairs, batch_size=20000)
     corrs = corrs.reshape(-1, n_replicates)
     null_dist = np.nanmedian(corrs, axis=1)
     return pd.Series(null_dist)
@@ -52,7 +52,7 @@ def corr_from_pairs(X: np.ndarray, pairs: dict, sameby: List[str]):
     list-like of correlation values and median of number of replicates
     '''
     pair_ix = np.vstack(list(pairs.values()))
-    corrs = corrcoef_indexed(X, pair_ix, batch_size=20000)
+    corrs = pairwise_corr(X, pair_ix, batch_size=20000)
     counts = [len(v) for v in pairs.values()]
 
     if len(sameby) == 1:
